+++ crates/client-utils/src/oracle/mod.rs
@@ -150,9 +150,6 @@ impl InMemoryOracle {
             println!("cycle-tracker-end: blob-verification");
             // kzg::verify_blob_kzg_proof(&blob.data, commitment, &blob.kzg_proof)
             // .map_err(|e| format!("blob verification failed for {:?}: {}", commitment, e))?;
-
-            // TODO: Would this allow us to leave 000...000 segments in blobs that were not empty and prove that?
-            // May need to track to ensure each blob element has been included.
         }
 
         Ok(())
