+++ crates/host-utils/Cargo.toml
@@ -16,3 +16,4 @@ sp1-sdk.workspace = true
 anyhow.workspace = true
 cargo_metadata.workspace = true
 serde_cbor.workspace = true
+revm.workspace = true
