+++ crates/host-utils/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod fetcher;
 pub mod helpers;
+pub mod precompile;
 
 use alloy_consensus::Header;
 use alloy_primitives::B256;
