+++ zkvm-host/bin/single.rs
@@ -2,13 +2,12 @@ use std::{env, fs};
 
 use anyhow::Result;
 use clap::Parser;
-use host_utils::{fetcher::SP1KonaDataFetcher, get_proof_stdin, ProgramType};
+use host_utils::{fetcher::SP1KonaDataFetcher, get_proof_stdin, precompile::precompile_hook, ProgramType};
 use kona_host::start_server_and_native_client;
 use num_format::{Locale, ToFormattedString};
 use sp1_sdk::{utils, ProverClient};
 
 use client_utils::precompiles::PRECOMPILE_HOOK_FD;
-use zkvm_host::precompile_hook;
 
 pub const SINGLE_BLOCK_ELF: &[u8] = include_bytes!("../../elf/zkvm-client-elf");
